Fix installment of jquery_easyXDM namespace
The provider did set a namespace (via easyXDM.noConflict("jquery_easyXDM")),
but it didn't actually create the namespace (global variable). When using
the FlashTransport protocol (e.g., IE6 and IE7), the Flash plugin would thus
try to evaluate JavaScript like "jquery_easyXDM.easyXDM.foo", which caused
a (silent) exception, ultimately causing easyXDM to fail to initialize.

This is the last of three fixes needed to get jquery.easyXDM to work on
IE6 and IE7. For debugging issues with the FlashTransport protocol in the
future, developers may find it helpful to force easyXDM to use Flash; this
makes it possible to debug the issue on more modern (and non-IE) browsers,
so long as they have a Flash plugin installed.
See "if (undef(protocol))" in prepareTransportStack();.