Show easyXDM debug output in-line in the test page
Create a "log" element on the consumer and provider pages if debugging
is enabled, so that debug output is sent there. This makes the easyXDM
debug output easily accessible on all browsers (e.g., no more annoying
pop-up windows in IE6 and IE7).

To make the provider page iframe visible, we pass a container element
id to easyXDM (jquery_easyXDM_provider_container). It's up to the
consumer to provide such an element; if easyXDM doesn't find the
element, the default behavior (placing the iframe off the page)
kicks in.